@@ -4,7 +4,7 @@ version: 1.1.0
 homepage: https://github.com/RatelHub/rflutter_alert
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
